+++ lib/duse/api/validations/base.rb
@@ -1,0 +1,14 @@
+module Duse
+  module API
+    module Validations
+      class Base
+        attr_reader :options
+
+        def initialize(options = {})
+          @options = options
+        end
+      end
+    end
+  end
+end
+
