--- lib/duse/api/validations/format_validation.rb
@@ -1,15 +1,0 @@
-require 'duse/api/validations/single_validation'
-
-class FormatValidation < SingleValidation
-  def invalid?(subject)
-    !subject.nil? && !subject.empty? && subject !~ format
-  end
-
-  def error_msg
-    "#{subject_name} contains illegal characters"
-  end
-
-  def format
-    options[:format]
-  end
-end
