--- lib/duse/api/validations/non_empty_validation.rb
@@ -1,11 +1,0 @@
-require 'duse/api/validations/single_validation'
-
-class NonEmptyValidation < SingleValidation
-  def invalid?(subject)
-    !subject.nil? && subject.empty?
-  end
-
-  def error_msg
-    "#{subject_name} must not be blank"
-  end
-end
