--- lib/duse/api/validations/single_validation.rb
@@ -1,12 +1,0 @@
-require 'duse/api/validations/validation'
-
-class SingleValidation < Validation
-  def validate(*subjects)
-    return [(options[:msg] || error_msg)] if invalid?(*subjects)
-    []
-  end
-
-  def subject_name
-    options[:subject_name]
-  end
-end
