--- lib/duse/api/validations/validation.rb
@@ -1,7 +1,0 @@
-class Validation
-  attr_reader :options
-
-  def initialize(options = {})
-    @options = options
-  end
-end
